+++ src/Stratis.Features.SystemContracts/SystemContractFeature.cs
@@ -20,6 +20,7 @@ using Stratis.Bitcoin.Features.SmartContracts.Rules;
 using Stratis.Bitcoin.Interfaces;
 using Stratis.Bitcoin.Utilities;
 using Stratis.Features.SystemContracts;
+using Stratis.Features.SystemContracts.Contracts;
 using Stratis.SmartContracts;
 using Stratis.SmartContracts.CLR;
 using Stratis.SmartContracts.CLR.Caching;
